+++ src/migration.py
@@ -65,6 +65,19 @@ def _validate_table_rows(
         current_group_level = row.level if row.is_group else row.level - 1
 
 
+def _generate_document_path(row: types_.TableRow) -> Path:
+    """Generate the path to a document from the table row.
+
+    Args:
+        row: Row containing link to document and path information.
+
+    Returns:
+        The path to the documentation file.
+
+    """
+    return Path(*row.path[:-1]) / f"{row.path[-1]}.md"
+
+
 def _create_document_meta(row: types_.TableRow) -> types_.DocumentMeta:
     """Create document meta file for migration from table row.
 
@@ -84,7 +97,7 @@ def _create_document_meta(row: types_.TableRow) -> types_.DocumentMeta:
             "Internal error, no implementation for creating document meta with missing link in row."
         )
     return types_.DocumentMeta(
-        path=Path(*row.path[:-1]) / f"{row.path[-1]}.md", link=row.navlink.link, table_row=row
+        path=_generate_document_path(row), link=row.navlink.link, table_row=row
     )
 
 
@@ -145,16 +158,63 @@ def _extract_docs_from_table_rows(
         yield _create_gitkeep_meta(row=previous_row)
 
 
-def _index_file_from_content(content: str) -> types_.IndexDocumentMeta:
+def _table_row_to_contents_index_line(row: types_.TableRow, discourse: Discourse) -> str:
+    """Generate a line of the contents index from a row of the navigation table.
+
+    Args:
+        row: the row of the navigation table.
+        discourse: Client to the documentation server.
+
+    Returns:
+        The contents index line.
+
+    """
+    leader = f"{(len(row.path) - 1) * '  '}1. "
+    if row.is_external(server_hostname=discourse.host):
+        return f"{leader}[{row.navlink.title}]({row.navlink.link})"
+    if row.is_group:
+        return f"{leader}[{row.navlink.title}]({Path(*row.path)})"
+    return f"{leader}[{row.navlink.title}]({_generate_document_path(row)})"
+
+
+def _migrate_navigation_table(rows: typing.Iterable[types_.TableRow], discourse: Discourse) -> str:
+    """Generate the contents index from the table rows of the navigation table.
+
+    Args:
+        rows: the rows of the navigation table.
+        discourse: Client to the documentation server.
+
+    Returns:
+        The contents index.
+
+    """
+    contents_index_lines = (
+        _table_row_to_contents_index_line(row=row, discourse=discourse) for row in rows
+    )
+    contents_index_body = "\n".join(contents_index_lines)
+    return f"# Contents\n\n{contents_index_body}".strip()
+
+
+def _index_file_from_content(
+    content: str,
+    table_rows: typing.Iterable[types_.TableRow],
+    discourse: Discourse,
+) -> types_.IndexDocumentMeta:
     """Get index file document metadata.
 
     Args:
         content: Index file content.
+        table_rows: Table rows from the index table.
+        discourse: Client to the documentation server.
 
     Returns:
         Index file document metadata.
     """
-    return types_.IndexDocumentMeta(path=Path("index.md"), content=content)
+    contents_index = _migrate_navigation_table(rows=table_rows, discourse=discourse)
+    # Strip any whitespace around file contents to avoid building up more and more whitespace
+    return types_.IndexDocumentMeta(
+        path=Path("index.md"), content=f"{content.strip()}\n\n{contents_index}"
+    )
 
 
 def make_parent(docs_path: Path, document_meta: types_.MigrationFileMeta) -> Path:
@@ -301,7 +361,10 @@ def _get_docs_metadata(
     Returns:
         Metadata of files to be migrated.
     """
-    index_doc = _index_file_from_content(content=index_content)
+    table_rows, table_rows_for_index = itertools.tee(table_rows, 2)
+    index_doc = _index_file_from_content(
+        content=index_content, table_rows=table_rows_for_index, discourse=discourse
+    )
     table_docs = _extract_docs_from_table_rows(table_rows=table_rows, discourse=discourse)
     return itertools.chain((index_doc,), table_docs)