+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Unreleased]
+
+## [v0.8.0] - 2023-11-30
+
+### Added
+
+- Support for migrating the navigation table to the contents index
+- Checks that external references on the contents index return a 2XX response to
+  HEAD requests
+
+### Fixed
+
+- Whitespace buildup between generated and input content
+
 ## [v0.7.0] - 2023-10-09
 
 ### Added